fix: update provider for jamfprotect-go-sdk v0.2.0 compatibility
Computer.MemorySize changed from *int64 to *float64 in the SDK. Update
the ComputerModel field, schema attribute, state builder, and test to
use Float64 throughout — no conversion needed and aligns with the SDK.

+++ internal/resources/computer/helpers_test.go
@@ -22,7 +22,7 @@ func TestBuildComputerModel_AllFields(t *testing.T) {
 	osPatch := int64(0)
 	arch := "arm64"
 	certID := "cert-123"
-	memSize := int64(17179869184) // 16 GB
+	memSize := float64(17179869184) // 16 GB
 	osString := "macOS 14.1"
 	kernelVer := "23.1.0"
 	installType := "pkg"
